backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
click==8.1.7
coloredlogs==15.0.1
coverage==7.6.10
cryptography==44.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
elastic-transport==8.17.0
elasticsearch==8.17.1
fastapi==0.115.6
fastapi-admin==1.0.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gkeepapi==0.15.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0rc1
gpsoauth==1.1.1
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.18
langchain-anthropic==0.3.7
langchain-community==0.3.17
langchain-core==0.3.34
langchain-deepseek==0.1.1
langchain-elasticsearch==0.3.2
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
langchain-openai==0.3.4
langchain-pinecone==0.2.2
langchain-postgres==0.0.13
langchain-redis==0.1.2
langchain-tests==0.3.11
langchain-text-splitters==0.3.6
langgraph==0.2.70
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.147
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.4
llama-index-readers-google==0.6.0
llama-index-readers-llama-parse==0.4.0
llama-index-readers-microsoft-onedrive==0.3.0
llama-index-readers-microsoft-outlook==0.3.0
llama-index-readers-web==0.3.5
llama-index-storage-chat-store-redis==0.4.0
llama-parse==0.6.0
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
ml-dtypes==0.4.1
mpmath==1.3.0
msal==1.31.1
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.4.7
openai==1.61.1
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pgvector==0.3.6
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
playwright==1.50.0
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic-settings==2.7.1
pydantic_core==2.27.1
PyDrive==1.3.1
pyee==12.1.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdf==5.3.0
pypika-tortoise==0.3.2
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-ulid==2.7.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
redisvl==0.3.9
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.28.1
sentence-transformers==3.4.1
sgmllib3k==1.0.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.36
sqlmodel==0.0.16
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
syrupy==4.8.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
time-machine==2.16.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tortoise-orm==0.22.2
tqdm==4.67.1
transformers==4.48.3
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
